refactor: move Art Institute of Chicago model files into its own package
