+++ src/main/java/uk/techreturners/VirtuArt/model/aicapi/AicArtworkResult.java
@@ -1,4 +1,4 @@
-package uk.techreturners.VirtuArt.model;
+package uk.techreturners.VirtuArt.model.aicapi;
 
 public record AicArtworkResult(AicArtwork data) {
 }
